+++ app/src/main/kotlin/org/fossify/gallery/fragments/VideoFragment.kt
@@ -18,6 +18,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.widget.ImageView
+import android.widget.RelativeLayout
 import android.widget.SeekBar
 import android.widget.TextView
 import androidx.appcompat.content.res.AppCompatResources
@@ -66,6 +67,7 @@ import org.fossify.gallery.activities.BaseViewerActivity
 import org.fossify.gallery.activities.VideoActivity
 import org.fossify.gallery.databinding.PagerVideoItemBinding
 import org.fossify.gallery.extensions.config
+import org.fossify.gallery.extensions.getActionBarHeight
 import org.fossify.gallery.extensions.getBottomActionsHeight
 import org.fossify.gallery.extensions.getFormattedDuration
 import org.fossify.gallery.extensions.getFriendlyMessage
@@ -91,6 +93,8 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
     companion object {
         private const val PROGRESS = "progress"
         private const val UPDATE_INTERVAL_MS = 250L
+        private const val TOUCH_HOLD_DURATION_MS = 300L
+        private const val TOUCH_HOLD_SPEED_MULTIPLIER = 2.0f
     }
 
     private var mIsFullscreen = false
@@ -118,6 +122,18 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
     private var mStoredBottomActions = true
     private var mStoredExtendedDetails = 0
     private var mStoredRememberLastVideoPosition = false
+    private var mOriginalPlaybackSpeed = 1f
+    private var mIsLongPressActive = false
+
+    private val mTouchHoldRunnable = Runnable {
+        mView.parent.requestDisallowInterceptTouchEvent(true)
+        // This code runs after the delay, only if the user is still holding down.
+        mIsLongPressActive = true
+        mOriginalPlaybackSpeed = mExoPlayer?.playbackParameters?.speed ?: mConfig.playbackSpeed
+        updatePlaybackSpeed(TOUCH_HOLD_SPEED_MULTIPLIER)
+
+        mPlaybackSpeedPill.fadeIn()
+    }
 
     private lateinit var mTimeHolder: View
     private lateinit var mBrightnessSideScroll: MediaSideScroll
@@ -130,6 +146,7 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
     private lateinit var mCurrTimeView: TextView
     private lateinit var mPlayPauseButton: ImageView
     private lateinit var mSeekBar: SeekBar
+    private lateinit var mPlaybackSpeedPill: TextView
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -170,6 +187,7 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
             }
 
             mSeekBar = bottomVideoTimeHolder.videoSeekbar
+            mPlaybackSpeedPill = playbackSpeedPill
             mSeekBar.setOnSeekBarChangeListener(this@VideoFragment)
             // adding an empty click listener just to avoid ripple animation at toggling fullscreen
             mSeekBar.setOnClickListener { }
@@ -178,6 +196,12 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
             mCurrTimeView = bottomVideoTimeHolder.videoCurrTime
             mBrightnessSideScroll = videoBrightnessController
             mVolumeSideScroll = videoVolumeController
+            mBrightnessSideScroll.onVerticalScroll = {
+                mTimerHandler.removeCallbacks(mTouchHoldRunnable)
+            }
+            mVolumeSideScroll.onVerticalScroll = {
+                mTimerHandler.removeCallbacks(mTouchHoldRunnable)
+            }
             mTextureView = videoSurface
             mTextureView.surfaceTextureListener = this@VideoFragment
 
@@ -215,6 +239,7 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
                 if (videoSurfaceFrame.controller.state.zoom == 1f) {
                     handleEvent(event)
                 }
+                handleTouchHoldEvent(event)
 
                 gestureDetector.onTouchEvent(event)
                 false
@@ -223,6 +248,15 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
 
         ViewCompat.setOnApplyWindowInsetsListener(binding.videoHolder) { _, insets ->
             val system = insets.getInsetsIgnoringVisibility(Type.systemBars())
+
+            val pillTopMargin = system.top + resources.getActionBarHeight(context) +
+                resources.getDimension(org.fossify.commons.R.dimen.normal_margin).toInt()
+            (mPlaybackSpeedPill.layoutParams as? RelativeLayout.LayoutParams)?.apply {
+                setMargins(
+                    0, pillTopMargin, 0, 0
+                )
+            }
+
             binding.bottomActionsDummy.updateLayoutParams<ViewGroup.LayoutParams> {
                 height = resources.getBottomActionsHeight() + system.bottom
             }
@@ -288,7 +322,6 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
                     doubleTap = { x, y ->
                         doSkip(false)
                     })
-
                 mVolumeSideScroll.initialize(
                     activity,
                     slideInfo,
@@ -940,6 +973,26 @@ class VideoFragment : ViewPagerFragment(), TextureView.SurfaceTextureListener,
                 height = screenHeight
             }
             mTextureView.layoutParams = this
+        }
+    }
+
+    private fun handleTouchHoldEvent(event: MotionEvent) {
+        when (event.action) {
+            MotionEvent.ACTION_DOWN -> {
+                if (mIsPlaying) {
+                    mTimerHandler.postDelayed(mTouchHoldRunnable, TOUCH_HOLD_DURATION_MS)
+                }
+            }
+
+            MotionEvent.ACTION_UP, MotionEvent.ACTION_CANCEL -> {
+                mTimerHandler.removeCallbacks(mTouchHoldRunnable)
+
+                if (mIsLongPressActive) {
+                    updatePlaybackSpeed(mOriginalPlaybackSpeed)
+                    mIsLongPressActive = false
+                    mPlaybackSpeedPill.fadeOut()
+                }
+            }
         }
     }
 }