+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Long press gesture to play videos at 2x speed ([#666])
+- 2x speed pill to indicate triggering of 2x playback speed gesture
 
 ## [1.9.1] - 2025-11-25
 ### Changed
@@ -246,7 +249,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#752]: https://github.com/FossifyOrg/Gallery/issues/752
 [#754]: https://github.com/FossifyOrg/Gallery/issues/754
 [#759]: https://github.com/FossifyOrg/Gallery/issues/759
-[#786]: https://github.com/FossifyOrg/Gallery/issues/786
 
 [Unreleased]: https://github.com/FossifyOrg/Gallery/compare/1.9.1...HEAD
 [1.9.1]: https://github.com/FossifyOrg/Gallery/compare/1.9.0...1.9.1